Close circuit breaker if the call on half-open state succeeds
According to the readme
```
Once the system has settled it will allow one call to go to the protected function. If the call succeds the breaker will be reset to its closed state otherwise it will continue open. This state is known as half open.
```
i.e. if the call performed in half open state succeed the circuit should
return to the original state.
Right now the circuit will continue in half-open state forever which
means that the circuit breaker is much more sensitive after opening
the first time, even if the failure that caused the circuit breaker
to open was completelly resolved.

This PR makes sure to reset the circuit breaker if the call in half-open
state has succeed.